loaded agendaitems in a more efficient place
