+++ app/components/agenda/agenda-header.hbs
@@ -48,8 +48,6 @@
             @didCloseMeeting={{@refreshRoute}}
             @onStartLoading={{this.setLoadingMessage}}
             @onStopLoading={{this.clearLoadingMessage}}
-            @notas={{@notas}}
-            @announcements={{@announcements}}
           />
         </Group.Item>
       {{/if}}
