+++ app/components/agenda/agenda-header/agenda-check.js
@@ -2,6 +2,7 @@ import Component from '@glimmer/component';
 import { A } from '@ember/array';
 import { trackedTask } from 'reactiveweb/ember-concurrency';
 import { task } from 'ember-concurrency';
+import CONSTANTS from 'frontend-kaleidos/config/constants';
 
 /**
  * @argument onSave
@@ -9,6 +10,27 @@ import { task } from 'ember-concurrency';
  */
 
 export default class AgendaHeaderAgendaCheck extends Component {
+
+  getAgendaitems = task(async () => {
+    const notas = [];
+    const announcements = [];
+    if (this.args.agenda) {
+      const agendaitems = await this.args.agenda.agendaitems;
+      const sortedAgendaitems = agendaitems?.slice().sort((a1, a2) => a1.number - a2.number);
+      for (const agendaitem of sortedAgendaitems) {
+        const type = await agendaitem.type;
+        if (type.uri === CONSTANTS.AGENDA_ITEM_TYPES.NOTA) {
+          notas.push(agendaitem);
+        } else {
+          announcements.push(agendaitem);
+        }
+      }
+    }
+    return { notas, announcements };
+  });
+
+  agendaitems = trackedTask(this, this.getAgendaitems);
+
   getFileNameMappings = task(async () => {
     const res = await fetch(`/document-naming/agenda/${this.args.agenda.id}`);
     const mappings = await res.json();
@@ -22,8 +44,8 @@ export default class AgendaHeaderAgendaCheck extends Component {
   fileNameMappings = trackedTask(this, this.getFileNameMappings);
 
   get notaGroups() {
-    const agendaitems = this.args.notas;
-    if (agendaitems.length > 0) {
+    const agendaitems = this.agendaitems.value?.notas;
+    if (agendaitems?.length > 0) {
       const mandatees = agendaitems.firstObject.mandatees;
       let currentSubmittersArray = mandatees.sortBy('priority');
       let currentItemArray = A([]);
