+++ app/components/agenda/agenda-header/agenda-version-actions.hbs
@@ -111,8 +111,6 @@
     @onCancel={{this.cancelApproveAgendaSidebar}}
     @onSave={{this.openConfirmApproveAgenda}}
     @saveLabel={{t "approve-design-agenda" serialnumber=@currentAgenda.serialnumber}}
-    @notas={{@notas}}
-    @announcements={{@announcements}}
   />
 {{/if}}
 
@@ -124,8 +122,6 @@
   @onCancel={{this.cancelApproveAgendaAndCloseMeetingSidebar}}
   @onSave={{this.openConfirmApproveAgendaAndCloseMeeting}}
   @saveLabel={{t "approve-agenda-and-close" serialnumber=@currentAgenda.serialnumber}}
-  @notas={{@notas}}
-  @announcements={{@announcements}}
 />
 {{/if}}
