+++ app/routes/agenda.js
@@ -1,14 +1,12 @@
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
 import { set } from '@ember/object';
-import CONSTANTS from 'frontend-kaleidos/config/constants';
 
 export default class AgendaRoute extends Route {
   @service('session') simpleAuthSession;
   @service agendaService;
   @service router;
   @service store;
-  @service throttledLoadingService;
 
   beforeModel(transition) {
     this.simpleAuthSession.requireAuthentication(transition, this.simpleAuthSession.unauthenticatedRouteName);
@@ -29,36 +27,10 @@ export default class AgendaRoute extends Route {
       include: 'status',
     });
 
-    const agendaitems = await this.store.query('agendaitem', {
-      filter: {
-        agenda: {
-          id: agenda.id,
-        },
-      },
-      include: 'mandatees',
-    });
-    const notas = [];
-    const announcements = [];
-    for (const agendaitem of agendaitems.sortBy('number').toArray()) {
-      const type = await agendaitem.type;
-      if (type.uri === CONSTANTS.AGENDA_ITEM_TYPES.NOTA) {
-        notas.push(agendaitem);
-      } else {
-        announcements.push(agendaitem);
-      }
-    }
-
-    // Load agendaitem pieces
-    await Promise.all(
-      agendaitems.map(this.throttledLoadingService.loadPieces.perform)
-    );
-
     return {
       meeting,
       agenda,
       reverseSortedAgendas,
-      notas,
-      announcements,
     };
   }
